pc_software/requirements.txt
pyserial
protobuf
pynput
fastapi
uvicorn[standard]
python-socketio
Jinja2
bleak
hidapi